+++ src/calendar.js
@@ -4,7 +4,7 @@
  * A javascript drag & drop event calendar, that is fully responsive and compatible with all modern browsers.
  * 
  * @file        calendar.js
- * @version     v2.1.6
+ * @version     v2.1.7
  * @author      Bunoon
  * @license     GNU AGPLv3
  * @copyright   Bunoon 2023
@@ -8974,8 +8974,17 @@ function calendarJs( elementOrId, options, searchOptions ) {
 
             contents.push( "BEGIN:VEVENT" );
             contents.push( "UID:" + getString( orderedEvent.id ) );
-            contents.push( "DTSTART:" + getICalDateTimeString( orderedEvent.from ) );
-            contents.push( "DTEND:" + getICalDateTimeString( orderedEvent.to ) );
+            contents.push( "STATUS:CONFIRMED" );
+            contents.push( "TRANSP:OPAQUE" );
+            contents.push( "SEQUENCE:0" );
+
+            if ( orderedEvent.isAllDay ) {
+                contents.push( "DTSTART:" + getICalDateString( orderedEvent.from ) );
+                contents.push( "DTEND:" + getICalDateString( orderedEvent.to ) );
+            } else {
+                contents.push( "DTSTART:" + getICalDateTimeString( orderedEvent.from ) );
+                contents.push( "DTEND:" + getICalDateTimeString( orderedEvent.to ) );
+            }
             
             if ( isDefinedDate( orderedEvent.created ) ) {
                 var created = getICalDateTimeString( orderedEvent.created );
@@ -9039,6 +9048,18 @@ function calendarJs( elementOrId, options, searchOptions ) {
             format.push( padNumber( eventDate.getHours() ) );
             format.push( padNumber( eventDate.getMinutes() ) );
             format.push( "00Z" );
+        }
+
+        return format.join( _string.empty );
+    }
+
+    function getICalDateString( eventDate ) {
+        var format = [];
+
+        if ( isDefined( eventDate ) ) {
+            format.push( eventDate.getFullYear() );
+            format.push( padNumber( eventDate.getMonth() + 1 ) );
+            format.push( padNumber( eventDate.getDate() ) );
         }
 
         return format.join( _string.empty );
@@ -10415,7 +10436,7 @@ function calendarJs( elementOrId, options, searchOptions ) {
      * @returns     {string}                                                The version number.
      */
     this.getVersion = function() {
-        return "2.1.6";
+        return "2.1.7";
     };
 
     /**