+++ docs/CHANGE_LOG.md
@@ -1,5 +1,11 @@
 # Calendar.js - Change Log:
 
+## Version 2.1.7:
+- Added correct support for All Day events for exported iCAL events (will show in the calendars correctly now).
+- Added "STATUS" (CONFIRMED), "TRANSP" (OPAQUE), and "STATUS" (0) support for exported iCAL events.
+
+<br>
+
 ## Version 2.1.6:
 - Improved the "PRODID" support for exported iCAL events.
 - Added "RRULE" support for exported iCAL events (states how an event should be repeated).